add interval time

+++ src/validation.cpp
@@ -8956,7 +8956,7 @@ void SelectMasterNode(unsigned int nCurrBlockHeight, uint32_t nTime)
     g_vecResultMasternodes.clear();
     std::vector<CMasternode>().swap(g_vecResultMasternodes);
     //1.3.3
-    g_nStartNewLoopTime = nTime*1000;
+    g_nStartNewLoopTime = (int64_t)nTime*1000;
     string strStartNewLoopTime = DateTimeStrFormat("%Y-%m-%d %H:%M:%S", g_nStartNewLoopTime/1000);
     string strBlockTime = DateTimeStrFormat("%Y-%m-%d %H:%M:%S", nTime);
     if(g_nStartNewLoopTime < nTime*1000)
